Move InvalidToken to jwt/core/
