+++ packages/jwt/src/core/invalid-token.error.ts
